# hw2/api/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: API сервиса маркетплейса
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Development server

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (Bearer <token>)

  schemas:
    # ==================== ENUMS ====================
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]

    UserRole:
      type: string
      enum: [USER, SELLER, ADMIN]

    DiscountType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]

    # ==================== ERROR ====================
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: Машиночитаемый код ошибки
          example: PRODUCT_NOT_FOUND
        message:
          type: string
          description: Человекочитаемое описание ошибки
        details:
          type: object
          nullable: true
          description: Дополнительные данные об ошибке
          additionalProperties: true

    # ==================== AUTH ====================
    RegisterRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
          maxLength: 128
        role:
          $ref: '#/components/schemas/UserRole'

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer

    UserResponse:
      type: object
      required: [id, username, role, created_at]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time

    # ==================== PRODUCT ====================
    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductListResponse:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # ==================== ORDER ====================
    OrderItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
          maximum: 999

    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
          minItems: 1
          maxItems: 50
        promo_code:
          type: string
          nullable: true
          pattern: '^[A-Z0-9_]{4,20}$'

    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
          minItems: 1
          maxItems: 50

    OrderItemResponse:
      type: object
      required: [id, product_id, quantity, price_at_order]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal

    OrderResponse:
      type: object
      required: [id, user_id, status, items, total_amount, discount_amount, created_at, updated_at]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        promo_code_id:
          type: integer
          nullable: true
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==================== PROMO CODE ====================
    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time

    PromoCodeResponse:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id:
          type: integer
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean

  # ==================== REUSABLE RESPONSES ====================
  responses:
    Unauthorized:
      description: Невалидный или просроченный токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

paths:
  # ==================== AUTH ====================
  /auth/register:
    post:
      tags: [auth]
      summary: Регистрация нового пользователя
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Имя пользователя уже занято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [auth]
      summary: Аутентификация, возвращает access + refresh token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Аутентификация успешна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновление access token по refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токен обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Невалидный или просроченный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== PRODUCTS ====================
  /products:
    get:
      tags: [products]
      summary: Список товаров с пагинацией и фильтрацией
      operationId: listProducts
      description: |
        Доступно всем аутентифицированным пользователям: USER, SELLER, ADMIN
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          schema:
            type: string
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [products]
      summary: Создание товара
      operationId: createProduct
      description: |
        SELLER — только свои товары. ADMIN — любые. USER — запрещено.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /products/{id}:
    get:
      tags: [products]
      summary: Получение товара по ID
      operationId: getProduct
      description: Доступно всем аутентифицированным пользователям
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные товара
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [products]
      summary: Обновление товара
      operationId: updateProduct
      description: |
        SELLER — только свои товары. ADMIN — любые. USER — запрещено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [products]
      summary: Мягкое удаление товара (перевод в ARCHIVED)
      operationId: deleteProduct
      description: |
        Выполняет мягкое удаление — переводит статус в ARCHIVED.
        SELLER — только свои товары. ADMIN — любые. USER — запрещено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Товар переведён в ARCHIVED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== ORDERS ====================
  /orders:
    post:
      tags: [orders]
      summary: Создание заказа
      operationId: createOrder
      description: |
        Доступно: USER, ADMIN. SELLER — запрещено.
        Проверяет rate limit, наличие активных заказов, наличие и статус товаров,
        остатки на складе, промокод (если передан).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            ACCESS_DENIED — недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: |
            PRODUCT_NOT_FOUND — товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            PRODUCT_INACTIVE — товар неактивен
            ORDER_HAS_ACTIVE — у пользователя уже есть активный заказ
            INSUFFICIENT_STOCK — недостаточно товара на складе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            PROMO_CODE_INVALID — промокод недействителен
            PROMO_CODE_MIN_AMOUNT — сумма заказа ниже минимальной
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: |
            ORDER_LIMIT_EXCEEDED — превышен лимит частоты создания заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}:
    get:
      tags: [orders]
      summary: Получение заказа по ID
      operationId: getOrder
      description: |
        USER — только свои заказы. ADMIN — любые. SELLER — запрещено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            ACCESS_DENIED — роль запрещена
            ORDER_OWNERSHIP_VIOLATION — заказ чужой
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ORDER_NOT_FOUND — заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [orders]
      summary: Обновление позиций заказа
      operationId: updateOrder
      description: |
        Только в состоянии CREATED. USER — только свои. ADMIN — любые. SELLER — запрещено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            ACCESS_DENIED или ORDER_OWNERSHIP_VIOLATION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ORDER_NOT_FOUND или PRODUCT_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            INVALID_STATE_TRANSITION, PRODUCT_INACTIVE, INSUFFICIENT_STOCK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: ORDER_LIMIT_EXCEEDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Отмена заказа
      operationId: cancelOrder
      description: |
        Отдельный эндпоинт для явной отмены.
        Разрешено только из состояний CREATED, PAYMENT_PENDING.
        USER — только свои. ADMIN — любые. SELLER — запрещено.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Заказ отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: ACCESS_DENIED или ORDER_OWNERSHIP_VIOLATION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_STATE_TRANSITION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== PROMO CODES ====================
  /promo-codes:
    post:
      tags: [promo-codes]
      summary: Создание промокода
      operationId: createPromoCode
      description: |
        Доступно: SELLER, ADMIN. USER — запрещено.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Промокод с таким кодом уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'